Create the BootstrapTrait
I added a test that installs Vanilla with the SiteTestTrait and found that a bunch of our other tests broke. After some digging I found that they were relying on a fresh bootstrap and container. To fix this I create a BootstrapTrait for tests to use if they want to test with the container.
