+++ tests/SiteTestTrait.php
@@ -14,10 +14,10 @@ use Garden\EventManager;
  * Allow a class to test against
  */
 trait SiteTestTrait {
-    /**
-     * @var Container
-     */
-    protected static $container;
+    use BootstrapTrait {
+        setupBeforeClass as private bootstrapBeforeClass;
+        teardownAfterClass as private bootstrapAfterClass;
+    }
 
     /**
      * @var array
@@ -30,25 +30,12 @@ trait SiteTestTrait {
     protected static $addons = ['vanilla', 'conversations', 'stubcontent'];
 
     /**
-     * Create the container for the site.
-     *
-     * @return Container Returns a container.
-     */
-    protected static function createContainer() {
-        $folder = strtolower(EventManager::classBasename(get_called_class()));
-        $bootstrap = new Bootstrap("http://vanilla.test/$folder");
-
-        $container = new Container();
-        $bootstrap->run($container);
-
-        return $container;
-    }
-
-    /**
      * Install the site.
      */
-    public static function setUpBeforeClass() {
-        $dic = self::$container = static::createContainer();
+    public static function setupBeforeClass() {
+        static::bootstrapBeforeClass();
+
+        $dic = self::$container;
 
         /* @var TestInstallModel $installer */
         $installer = $dic->get(TestInstallModel::class);
@@ -70,15 +57,6 @@ trait SiteTestTrait {
      */
     public static function teardownAfterClass() {
         self::$addons = ['vanilla', 'conversations', 'stubcontent'];
-        Bootstrap::cleanup(self::container());
-    }
-
-    /**
-     * Get the container for the site info.
-     *
-     * @return Container Returns a container with site dependencies.
-     */
-    protected static function container() {
-        return self::$container;
+        static::bootstrapAfterClass();
     }
 }
